requirements.txt
# Plant AI System Requirements
# Core ML/AI Libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
sentence-transformers>=2.2.0
clip-by-openai>=1.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0
imgaug>=0.4.0

# Object Detection and Segmentation
ultralytics>=8.0.0  # YOLOv8
detectron2>=0.6.0   # Mask R-CNN
mmcv>=2.0.0
mmdet>=3.0.0

# Data Processing and Analysis
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Web Framework
Flask>=2.3.0
Flask-CORS>=4.0.0

# OpenAI API
openai>=1.0.0

# Utilities
tqdm>=4.65.0
psutil>=5.9.0
python-dotenv>=1.0.0
wandb>=0.15.0  # For experiment tracking
tensorboard>=2.13.0

# Data Loading and Augmentation
torchvision>=0.15.0
torchaudio>=2.0.0